requirements.txt
# base
Django>=5.2,<5.3
whitenoise>=6.11.0
gunicorn>=24.1.1
django-cors-headers>=4.9.0
python-dotenv>=1.0.1

# db
psycopg>=3.2.3

# images
pillow>=11.1.0  

# drf & jwt
djangorestframework>=3.16.1
django-filter>=24.3

# testing
selenium>=4.40.0

# admin
django-unfold==0.77.1
django-solo>=2.3.0

# tools
requests>=2.32.3
